+++ Karta/src/main/java/app/repository/KartaRepository.java
@@ -9,15 +9,15 @@ import org.springframework.stereotype.Repository;
 import app.entities.Karta;
 
 @Repository
-public interface KartaRepository extends JpaRepository<Karta, Integer>{
+public interface KartaRepository extends JpaRepository<Karta, Long>{
 
 	//Selekcija karte po idLet-a
 	@Query("select k from Karta k where k.idLeta like :idLeta")
-	Karta selectCardByPlaneID(int idLeta);
+	Karta selectCardByPlaneID(long idLeta);
 	
-	boolean existsByIdLeta(int idLeta);
+	boolean existsByIdLeta(long idLeta);
 	
 	//Selekcija svih idUsera koji su kupili kartu za odredjeni let
 	@Query("select k.idUser from Karta k where k.idLeta like :idLeta")
-	List<Integer> findByIdUseraByIdLeta(int idLeta);
+	List<Long> findByIdUseraByIdLeta(long idLeta);
 }
