+++ Let/src/main/java/app/repository/AvionRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import app.entities.Avion;
 
 @Repository
-public interface AvionRepository extends JpaRepository<Avion, Integer>{
+public interface AvionRepository extends JpaRepository<Avion, Long>{
 
 	Avion findByIdAviona(int idAviona);
