+++ Registrator/src/main/java/app/repository/User_KarticaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import app.entities.User_Kartica;
 
 @Repository
-public interface User_KarticaRepository extends JpaRepository<User_Kartica, Integer> {
+public interface User_KarticaRepository extends JpaRepository<User_Kartica, Long> {
 
 }
